feat(server): adding arguments parser

+++ lib/log.c
@@ -12,10 +12,10 @@
     va_end(args);        \
     printf("\n");
 
-uint8_t log__timestamps = 0;
-uint8_t log__stack_trace = 0;
-uint8_t log__enable = 1;
-uint8_t log__frame = 0;
+int log__timestamps = 0;
+int log__stack_trace = 0;
+int log__enable = 1;
+int log__frame = 0;
 
 static inline void print_label(const char* label) {
     if (log__timestamps) {
