+++ lib/log.h
@@ -11,10 +11,10 @@
  */
 
 /** log options **/
-extern uint8_t log__timestamps;
-extern uint8_t log__stack_trace;
-extern uint8_t log__enable;
-extern uint8_t log__frame;
+extern int log__timestamps;
+extern int log__stack_trace;
+extern int log__enable;
+extern int log__frame;
 
 void _log__print(const char* label, const char* fmt, ...);
 void _log__info(const char* fmt, ...);
